refactor: consolidate duplicated code in impact analysis clients and workflows
- Lifted environment variables to job level in `ci.yml`.
- Created `scripts/lib/codeReviewUtils.ts` with shared code review logic.
- Refactored `geminiCodeReviewClient.ts` and `githubModelsCodeReviewClient.ts` to use shared utilities.
- Consolidated `parseVisualReviewFindings` in `scripts/lib/visualReviewUtils.ts`.
- Updated visual review clients to use the shared parser.
